Separated and organized files and deps
